+++ doc/release-notes.md
@@ -11,3 +11,13 @@ This release includes the following features and fixes:
    until v0.21 at which time `getaccountaddress` will also be removed.  To
    use `getaccountaddress` start `bitcoind` with the `-deprecatedrpc=accounts`
    option.  See the v0.20.6 release notes for more details.
+
+Network
+-------
+ - When fetching a transaction announced by multiple peers, previous versions of
+   Bitcoin ABC would sequentially attempt to download the transaction from each
+   announcing peer until the transaction is received, in the order that those
+   peers' announcements were received.  In this release, the download logic has
+   changed to randomize the fetch order across peers and to prefer sending
+   download requests to outbound peers over inbound peers. This fixes an issue
+   where inbound peers can prevent a node from getting a transaction.